third_party/blink/web_tests/external/wpt/html/dom/documents/dom-tree-accessors/nameditem-names-expected.txt

This is a testharness.js-based test.
[FAIL] An embed name does not appears in a document's property names if the embed is inside another embed.
  assert_false: expected false got true
[FAIL] An object name appears in a document's property names if the object is exposed.
  assert_true: expected true got false
[FAIL] An object name does not appear in a document's property names if the object is inside another object.
  assert_false: expected false got true
[FAIL] An object id does not appear in a document's property names if the object is inside another object.
  assert_false: expected false got true
[FAIL] A document's property names appear in tree order.
  assert_greater_than: expected a number greater than 11 but got 0
Harness: the test ran to completion.
